+++ imports/util/validate/zipcode.js
@@ -1,0 +1,12 @@
+const zipCode = value => {
+  const regex = /(^\d{5}$)|(^\d{5}-\d{4}$)/;
+  return regex.test(value);
+};
+
+const locationBasedZipCode = value => {
+  if (value === "Using your location") return true;
+  const regex = /(^\d{5}$)|(^\d{5}-\d{4}$)/;
+  return regex.test(value);
+};
+
+export { zipCode, locationBasedZipCode };